LayoutTests/fast/text/content-language-dynamically-added.html

<!DOCTYPE html>
<html>
<head id="head">
<link rel="stylesheet" href="../js/resources/js-test-style.css">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<p>Test for <a href="https://bugs.webkit.org/show_bug.cgi?id=76701">bug 76701</a>:
map http-equiv content-language to -webkit-locale. This particular test tests
that a dynamically added content-language is processed. This expectation may
change, see bug. The HTML 5 spec decrees that http-equiv content-language be
processed when it is
<a href="http://dev.w3.org/html5/spec/Overview.html#insert-an-element-into-a-document">inserted into the document</a>.
In Firefox and IE adding the meta element dynamically doesn't seem to affect
the page, but in Firefox the meta content-language displayed in "Page
Information" is affected.
</p>
<div id="console"></div>
<div id="x"></div>
<div id="y" lang="ar"></div>
<script>
function languageOfNode(id) {
    var element = document.getElementById(id);
    return window.getComputedStyle(element).webkitLocale;
}
shouldBeEqualToString("languageOfNode('x')", "auto");
shouldBeEqualToString("languageOfNode('y')", "ar");

var meta = document.createElement("meta");
meta.httpEquiv = "content-language";
meta.content = "ja";
document.getElementById("head").appendChild(meta);
shouldBeEqualToString("languageOfNode('x')", "ja");
shouldBeEqualToString("languageOfNode('y')", "ar");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
